chore: refactor routing tests
